mcp_servers/aws_api_mcp/requirements.txt
awslabs.aws-api-mcp-server>=1.0.0
boto3>=1.35.0
